# openapi.yaml
openapi: 3.0.0
info:
  title: ZENU - AI Agent for Radio Presenters API
  description: |
    ZENU is an AI-powered digital producer for radio presenters and OAPs.
    This API provides digital skills for IBM watsonx Orchestrate integration.
    
    **Core Workflow:**
    Fetch News → Clean & Summarize → Rewrite in OAP Tone → Build Script → Deliver to Teleprompter
  version: 1.0.0
  contact:
    name: ZENU Support
    email: support@zenu.app
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://zenu.app/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new presenter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "DJ Goodvibes"
                email:
                  type: string
                  format: email
                  example: "dj@radio.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login presenter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  # ==================== DIGITAL SKILL 1: FETCH NEWS ====================
...
  /profile:
    get:
      tags:
        - Profile
      summary: Get presenter profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenterProfile'
    
    put:
      tags:
        - Profile
      summary: Update presenter profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresenterProfile'
      responses:
        '200':
          description: Profile updated successfully

  # ==================== COMPLETE WORKFLOW ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
...
    PresenterProfile:
      type: object
      properties:
        userId:
          type: string
        preferredLanguage:
          type: array
          items:
            type: string
          example: ["English", "Pidgin"]
        speakingSpeed:
          type: string
          enum: [slow, medium, fast]
        signatureIntro:
          type: string
        signatureOutro:
          type: string
        topicPreferences:
          type: array
          items:
            type: string
        showStructure:
          type: array
          items:
            type: object
            properties:
              section:
                type: string
              duration:
                type: integer
        voiceSample:
          type: string
          format: uri